requirements.txt
# CADI - Climate-Aware Deal Intelligence
# Core dependencies

# LLM / Agent
anthropic>=0.40.0

# PDF parsing
pymupdf>=1.24.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
folium>=0.16.0

# NER (optional - Claude can do NER via prompting)
# spacy>=3.7.0

# Report generation
python-docx>=1.1.0
jinja2>=3.1.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0  # Pretty terminal output for demo
